feat: Adds confirmation Dialog

+++ src/components/data-table.tsx
@@ -3,6 +3,7 @@
 import * as React from "react";
 import { Trash } from "lucide-react";
 
+import { useConfirm } from "@/hooks/use-confirm";
 import {
   ColumnDef,
   ColumnFiltersState,
@@ -43,6 +44,11 @@ export function DataTable<TData, TValue>({
   onDelete,
   disabled,
 }: DataTableProps<TData, TValue>) {
+  const [ConfirmDialog, confirm] = useConfirm(
+    "Are you Sure?",
+    "You are about to delete the selected rows"
+  );
+
   const [sorting, setSorting] = React.useState<SortingState>([]);
   const [columnFilters, setColumnFilters] = React.useState<ColumnFiltersState>(
     []
@@ -68,6 +74,7 @@ export function DataTable<TData, TValue>({
 
   return (
     <div>
+      <ConfirmDialog />
       <div className="flex items-center py-4">
         <Input
           placeholder={`Filter ${filterKey}...`}
@@ -83,9 +90,12 @@ export function DataTable<TData, TValue>({
             variant="outline"
             size="sm"
             disabled={disabled}
-            onClick={() => {
-              onDelete(table.getFilteredSelectedRowModel().rows);
-              table.resetRowSelection();
+            onClick={async () => {
+              const ok = await confirm();
+              if (ok) {
+                onDelete(table.getFilteredSelectedRowModel().rows);
+                table.resetRowSelection();
+              }
             }}
           >
             <Trash className="size-4 mr-2" />
